% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_data}
\alias{check_data}
\title{check_order}
\usage{
check_data(
  data,
  order_continuous,
  formula_continuous,
  formula_discrete,
  n,
  t,
  state_varying_continuous = c(),
  state_varying_discrete = list()
)
}
\arguments{
\item{data}{Included data frame}

\item{order_continuous}{Vector of parameter names declared to be continuous}
}
\description{
check_order
}
