% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_tp_s}
\alias{check_tp_s}
\title{check_tp_s}
\usage{
check_tp_s(shared_TP = NULL, shared_S = NULL, n = NULL)
}
\arguments{
\item{shared_TP}{TRUE/FALSE indicating whether transition probabilities are being shared.}

\item{shared_S}{TRUE/FALSE indicating whether the discrete process is being shared

This function checks for the case that the user indicates that transition probabilities are individual but the state process is shared
which is not possible. Returns an error.}
}
\description{
check_tp_s
}
