% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_order_vector}
\alias{create_order_vector}
\title{create_order_vector}
\usage{
create_order_vector(formula, order_continuous)
}
\arguments{
\item{formula}{The formula list}

\item{order_continuous}{A character vector indicating which parameters are ordered.

This function creates a 0/1 vector indicating whether a particular parameter that is varying across states is ordered.}
}
\description{
create_order_vector
}
