% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{deparse_call_formula}
\alias{deparse_call_formula}
\title{deparse_call_formula}
\usage{
deparse_call_formula(x)
}
\arguments{
\item{x}{A language object retrieved from call}
}
\description{
deparse_call_formula
}
