% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_nt_k2.R, R/rstanmsm_dgp.R
\name{dgp_nt_k2}
\alias{dgp_nt_k2}
\title{DGP for MSM-TVTP balanced panel}
\usage{
dgp_nt_k2(N, T, Mx_var, Mx_sha, Mz_var, Mz_sha, Mx_var_int = FALSE,
  Mx_sha_int = FALSE, Mz_var_int = FALSE, Mz_sha_int = FALSE,
  para.var = FALSE, gamma = NULL, lambda = NULL, zeta = NULL,
  beta = NULL, fit = FALSE)

dgp_nt_k2(N, T, Mx_var, Mx_sha, Mz_var, Mz_sha, Mx_var_int = FALSE,
  Mx_sha_int = FALSE, Mz_var_int = FALSE, Mz_sha_int = FALSE,
  para.var = FALSE, gamma = NULL, lambda = NULL, zeta = NULL,
  beta = NULL, fit = FALSE)
}
\arguments{
\item{N}{The number of separate time series.}

\item{T}{The length of each time series.}

\item{gamma}{A length 2 vector of intercepts for the TVTP.}

\item{lambda}{A length 2 vector of coefficients on the predictor on z for the TVTP.}

\item{mu}{A length 2 vector of intercepts for the linear models.}

\item{phi}{A length 2 vector of AR1 coefficients. Each should be between 0 and 1.}

\item{N}{The number of separate time series.}

\item{T}{The length of each time series.}

\item{gamma}{A length 2 vector of intercepts for the TVTP.}

\item{lambda}{A length 2 vector of coefficients on the predictor on z for the TVTP.}

\item{mu}{A length 2 vector of intercepts for the linear models.}

\item{phi}{A length 2 vector of AR1 coefficients. Each should be between 0 and 1.}
}
\value{
dgp_nt() returns three lists. The first lists contains the observed data.
The second list contains the parameters.
The third list contains summary statistics.
\describe{
  \item{switches}{The total number of switches of all discrete processes.}
  \item{t.1}{The total time spent in state 1.}
  \item{t.2}{The total time spent in state 1.}
}

dgp_nt() returns three lists. The first lists contains the observed data.
The second list contains the parameters.
The third list contains summary statistics.
\describe{
  \item{switches}{The total number of switches of all discrete processes.}
  \item{t.1}{The total time spent in state 1.}
  \item{t.2}{The total time spent in state 1.}
}
}
\description{
Generate data for a MSM-TVTP balanced panel

Generate data for a MSM-TVTP balanced panel
}
\details{
\code{dgp_nt} simulates data to test the model. It outputs the observable data,
the parameters used to generate the data, and summary statistics. The summary statistics
include the number of switches between states and the time spent in each state.

\code{dgp_nt} simulates data to test the model. It outputs the observable data,
the parameters used to generate the data, and summary statistics. The summary statistics
include the number of switches between states and the time spent in each state.
}
\examples{
dgp_nt(N = 2, T = 300, gamma = c(0.25, 0.5), lambda = c(0.75, 0.5), mu = c(2, 6), phi = c(0.2, 0.8))

dgp_nt(N = 2, T = 300, gamma = c(0.25, 0.5), lambda = c(0.75, 0.5), mu = c(2, 6), phi = c(0.2, 0.8))
}
\keyword{DGP}
