% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.add_intercept_name}
\alias{.add_intercept_name}
\title{.add_intercept_name}
\usage{
.add_intercept_name(has_intercept, names)
}
\arguments{
\item{has_intercept}{Flag indicating whether the parameter set has an intercept.}

\item{names}{Names of the parameters}
}
\value{
Return the names of the parameters with `Intercept` appended at the front.
}
\description{
.add_intercept_name
}
