% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.create_predmat}
\alias{.create_predmat}
\title{.create_predmat}
\usage{
.create_predmat(names, data)
}
\arguments{
\item{names}{Predictor names}

\item{data}{The dataframe}
}
\description{
.create_predmat
}
\details{
Based on the parsed formula adjusts the data frame
}
