% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.factor_mat}
\alias{.factor_mat}
\title{.factor_mat}
\usage{
.factor_mat(x, name)
}
\arguments{
\item{name}{}
}
\description{
.factor_mat
}
\examples{
N <- 20
x <- as.factor(sample(seq(1, 4), N, replace = TRUE))
.factor_mat(x, "NAME")
}
