% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.interaction}
\alias{.interaction}
\title{.interaction}
\usage{
.interaction(x, data)
}
\arguments{
\item{x}{The variable name}

\item{data}{The data}
}
\description{
.interaction
}
\examples{
N <- 20
x <- "x1:x2"
data <- data.frame(x1 = as.factor(sample(seq(1, 4), N, replace = TRUE)), x2 = rnorm(N))
.interaction(x = x, data = data)
}
