% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.j_define}
\alias{.j_define}
\title{.j_define}
\usage{
.j_define(data = data, j = j, n = n)
}
\arguments{
\item{j}{Grouping variable}

\item{n}{Unit variable}
}
\description{
.j_define
}
\details{
j Can be either individual specific (everyone has their own state process),
shared (everyone has the same state process) or based on indicators.
}
\examples{
N <- 50
j <- "individual"
n <- sample(seq(1, 4), N, replace = TRUE)
.j_define(j = j, n = n)

j <- "shared"
n <- sample(seq(1, 4), N, replace = TRUE)
.j_define(j = j, n = n)

j <- sample(seq(1,2), N, replace = TRUE)
n <- sample(seq(1, 4), N, replace = TRUE)
.j_define(j = j, n = n)
}
