% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.n_adj}
\alias{.n_adj}
\title{.n_adj}
\usage{
.n_adj(n = NULL, data)
}
\arguments{
\item{n}{A vector of length N indicating group association.}

\item{data}{The data vector}
}
\value{
Returns a length n vector of 1, ..., J flags.
}
\description{
.n_adj
}
