% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.sort_and_na_extend}
\alias{.sort_and_na_extend}
\title{.na_extend}
\usage{
.sort_and_na_extend(data, n = NULL, t = NULL, j = NULL)
}
\arguments{
\item{n}{The unit level indicator}

\item{t}{The timepoint indicator}

\item{j}{The group level indicator}

\item{df}{A dataframe containing n and z}
}
\description{
.na_extend
}
\examples{
obs <- 200
N <- 5
J <- 3
data <- data.frame(x = rnorm(obs), n = sort(rep(seq(1, N), obs/N)), t = rep(seq(1, obs/N), N))
data <- merge(data, data.frame(n = seq(1, N), j = sample(seq(1, J), N, replace = T)))
data <- data[sample(seq(1,obs), round(0.8 * obs)) , ]
out <- .sort_and_na_extend(data = data, n = "n", j = "j", t = "t")
}
