% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.split_formula}
\alias{.split_formula}
\title{.split_formula}
\usage{
.split_formula(x)
}
\arguments{
\item{x}{The formula}
}
\description{
.split_formula
}
\details{
1.
}
\examples{
x <- "y ~ 1 + x1#x2"
.split_formula(x)

x <- "y ~ x1:x2"
.split_formula(x)
}
