% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.state_probabilities}
\alias{.state_probabilities}
\title{.state_probabilities}
\usage{
.state_probabilities(data = data, j = j, q = q)
}
\arguments{
\item{j}{The state process indicator.}

\item{q}{The transition probability indicator.}
}
\description{
.state_probabilities
}
\details{
This function creates a vector of length J assigning the state
processes to specific transition probabilities.
}
\examples{
t <- data.frame(j = seq(1, 4), q = rep(seq(1, 2), 2))
N <- 20
data <- t[sample(seq(1, 4), N, replace = TRUE), ]
.state_probabilities(data = data, j = "j", q = "q")
.state_probabilities(data = data, j = "j", q = NULL)
.state_probabilities(data = data, j = "j", q = "individual")
.state_probabilities(data = data, j = "j", q = "shared")
}
