% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{formula_parse}
\alias{formula_parse}
\title{formula_parse}
\usage{
formula_parse(formula_discrete = NULL,
  formula_continuous = formula_continuous)
}
\description{
Parses the formulas into their components and stores them in a list
}
