% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_msm.dgp.R
\name{msm_dgp}
\alias{msm_dgp}
\title{msm_dgp}
\usage{
msm_dgp(N, T, K, S, N_tp, p)
}
\arguments{
\item{N}{Number of units}

\item{T}{Number of average timepoints}

\item{K}{Number of states}

\item{S}{Number of maximum state processes}

\item{N_tp}{Number of transition probabilities}

\item{p}{Probability that an observation is missing.}
}
\description{
msm_dgp
}
