% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{naming_fun}
\alias{naming_fun}
\title{naming_fun}
\usage{
naming_fun(x, para_names)
}
\arguments{
\item{x}{is a vector of names}

\item{para_names}{is a list of objects with names}
}
\description{
naming_fun
}
