% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prepare_standata}
\alias{prepare_standata}
\title{prepare_standata}
\usage{
prepare_standata(
  data. = data,
  priors. = priors,
  order_continuous. = order_continuous,
  state_sigma. = state_sigma,
  tvtp. = 0
)
}
\description{
order_continuous <- c("Intercept")
priors <- prior_mat(prior = NULL, K = 3, y = data$data_lst$y)
prepare_standata(data = data$data_lst, priors = priors,
                 order_continuous = order_continuous, state_sigma = FALSE, tvtp = TRUE)
}
