% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.stan_msm}
\alias{print.stan_msm}
\title{print.stan_msm}
\usage{
\method{print}{stan_msm}(object, digits = 1, detail = TRUE, ...)
}
\arguments{
\item{object}{A stan_msm fit object.}

\item{digits}{The number of digits to round to.}

\item{detail}{Print output.

A S3 function to print a summary output of a stan_msm object.}
}
\description{
print.stan_msm
}
