% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_mat}
\alias{prior_mat}
\title{prior_mat}
\usage{
prior_mat(prior = prior, K = K, y = y)
}
\arguments{
\item{prior:}{A list of the priors.}

\item{K:}{The number of states}

\item{y:}{The outcome}
}
\value{
Returns a list with the transition probability matrix Dirichlet parameters and a matrix of the priors for the rest.
}
\description{
prior_mat
}
\details{
If priors are null, then everything becomes normal.
}
