% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_msm.fit.R
\name{rstan_msm.fit}
\alias{rstan_msm.fit}
\title{rstan_msm_fit: Call for rstan_msm}
\usage{
rstan_msm.fit(x_var = x_var, x_fix = x_fix, z_var = z_var,
  z_fix = z_fix, y = y, N = n, T = t, K = 2,
  family = gaussian())
}
\description{
rstan_msm_fit: Call for rstan_msm
}
