% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanmsm.R
\docType{summary}
\name{rstanmsm-package}
\alias{rstanmsm-package}
\title{rstanmsm: A package for Bayesian Inference for Markov Switching Models with Time-Varying Transition Probabilities}
\description{
The package provides an easily accessible format to estimate MSM models with TVTP and for TSCS data.
}
\section{Functions}{
NA
}

