% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{slicer_time}
\alias{slicer_time}
\title{slicer_time}
\usage{
slicer_time(j)
}
\arguments{
\item{j}{The group indicator}
}
\description{
slicer_time
}
\details{
Returns a matrix that contains the start and stop for each state process
and further grabs elements from the matrix that contains the start and stop of individual
time points
}
\examples{
j <- sort(rep(seq(1, 10), 20))
slicer_time(j = j)
}
