% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{split_naming}
\alias{split_naming}
\title{split_naming}
\usage{
split_naming(x, names_list, N, K, shared_TP = TRUE)
}
\arguments{
\item{x}{A named vector of coefficients or standard errors}
}
\description{
split_naming
}
