% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_msm.fit.R
\name{stan_msm.fit}
\alias{stan_msm.fit}
\title{stan_msm.fit}
\usage{
stan_msm.fit(
  data = data,
  state_sigma = FALSE,
  order_continuous = order_continuous,
  priors = priors,
  family = gaussian(),
  init_prior = FALSE,
  algorithm = c("optimizing", "sampling"),
  ...
)
}
\arguments{
\item{data}{Data is a list including matrixes for the sorted predictors and indicator vectors for n and t}

\item{state_sigma}{text}

\item{order_continuous}{text}

\item{priors}{A list object containing the prior for the Dirichlet and all other parameters.}
}
\description{
stan_msm.fit
}
