% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{start_stop_slicer}
\alias{start_stop_slicer}
\title{start_stop_slicer}
\usage{
start_stop_slicer(j = j, t = t)
}
\description{
start_stop_slicer
}
\examples{
N <- 360
j <- sample(seq(1, 3), N, replace = TRUE)
t <- sort(sample(rep(seq(1, 60), 6)))
start_stop_slicer(j = j, t = t)
}
