% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{data_parse}
\alias{data_parse}
\title{data_parse}
\usage{
data_parse(formula_continuous, formula_discrete, state_varying_continuous,
  state_varying_discrete, data, n_var, t_var, K)
}
\description{
data_parse
}
