% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_msm.fit.R
\name{rstan_msm.fit}
\alias{rstan_msm.fit}
\title{rstan_msm_fit: Call for rstan_msm}
\usage{
rstan_msm.fit(x_var, x_sha, z_var, z_sha, y = y, n, t, K = 2,
  family = gaussian(), nchains = 4, niter = 2000,
  warmup = floor(niter/2), thin = 1, controls = NULL,
  init.prior = FALSE)
}
\description{
rstan_msm_fit: Call for rstan_msm
}
