% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_msm.fit.R
\name{stan_msm.fit}
\alias{stan_msm.fit}
\title{stan_msm.fit}
\usage{
stan_msm.fit(data, K = 2, shared_TP = TRUE, shared_S = FALSE,
  state_SD = FALSE, order_continuous, family = gaussian(),
  init.prior = TRUE, algorithm = c("optimizing", "sampling"),
  ... = ...)
}
\arguments{
\item{data}{Data is a list including matrixes for the sorted predictors and indicator vectors for n and t}
}
\description{
stan_msm.fit
}
